Initial break up
